--- index.js
@@ -1,5 +1,0 @@
-// filepath: /C:/Users/Aluno/Documents/GitHub/teste/index.js
-import { registerRootComponent } from 'expo';
-import Navigation from './Navigation';
-
-registerRootComponent(Navigation);
