--- navigation.js
@@ -1,21 +1,0 @@
-// filepath: /C:/Users/Aluno/Documents/GitHub/teste/Navigation.js
-import * as React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import App from './App';
-import SecondScreen from './SecondScreen';
-
-const Stack = createStackNavigator();
-
-function Navigation() {
-  return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName="Home">
-        <Stack.Screen name="Home" component={App} />
-        <Stack.Screen name="SecondScreen" component={SecondScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
-}
-
-export default Navigation;
