+++ package.json
@@ -9,18 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~3.1.3",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/stack": "^7.1.1",
-    "expo": "~50.0.14",
+    "expo": "~50.0.17",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
     "react-native": "0.73.6",
-    "react-native-paper": "^5.13.1",
-    "react-native-safe-area-context": "^5.2.0",
-    "react-native-screens": "^4.6.0",
-    "react-native-web": "~0.19.6"
+    "react-native-web": "~0.19.6",
+    "react-dom": "18.2.0",
+    "@expo/metro-runtime": "~3.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
